fix: perbaiki masalah jenjang dimana masih statis sd, smp dan sma, menjadi lebih luas dengan prefix

+++ config/absensi.php
@@ -10,5 +10,5 @@ return [
     | otomatis dikunci. Guru masih bisa edit absensi selama masa tenggang ini.
     |
     */
-    'auto_lock_grace_minutes' => env('ABSENSI_LOCK_GRACE', 30),
+    'auto_lock_grace_minutes' => (int) env('ABSENSI_LOCK_GRACE', 30),
 ];
